refactor: move 'containers' into 'publishing' app
