+++ openedx_learning/api/authoring_models.py
@@ -9,7 +9,6 @@ consistent.
 # pylint: disable=wildcard-import
 from ..apps.authoring.collections.models import *
 from ..apps.authoring.components.models import *
-from ..apps.authoring.containers.models import *
 from ..apps.authoring.contents.models import *
 from ..apps.authoring.publishing.model_mixins import *
 from ..apps.authoring.publishing.models import *
